requirements.txt
aioredis==2.0.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
click==8.1.2
dnspython==2.2.1
email-validator==1.2.0
fastapi==0.75.2
fastapi-jwt-auth==0.5.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-resources==5.7.1
jsonschema==4.5.1
motor==2.5.1
protobuf==3.20.1
pydantic==1.9.0
PyJWT==1.7.1
pymongo==3.12.3
pyrsistent==0.18.1
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
